backend/requirements.txt
fastapi
uvicorn
python-multipart
chromadb
pypdf
python-docx
langchain
langchain-community
langchain-core
